Updated web.xml and add missing student crud files

+++ dmit2015-jpa-faces-app/src/main/java/dmit2015/service/JakartaPersistenceTaskService.java
@@ -8,7 +8,6 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.security.enterprise.SecurityContext;
 import jakarta.transaction.Transactional;
-import jakarta.validation.Valid;
 
 import java.util.List;
 import java.util.Optional;
@@ -27,16 +26,16 @@ public class JakartaPersistenceTaskService implements TaskService {
 
     @Override
     @Transactional
-    public Task createTask(@Valid Task task) {
+    public Task createTask(Task task) {
         // If the primary key is not an identity column then write code below here to
         // 1) Generate a new primary key value
         // 2) Set the primary key value for the new entity
-        task.setId(UUID.randomUUID().toString());
         String username = _securityContext.getCallerPrincipal().getName();
         if (username.equalsIgnoreCase("anonymous")) {
-            throw new RuntimeException("Access denied. Anonymous users are not allowed.");
+            throw new SecurityException("Access denied. Anonymous users are not allowed.");
         }
         task.setUsername(username);
+        task.setId(UUID.randomUUID().toString());
         entityManager.persist(task);
         return task;
     }
@@ -57,23 +56,30 @@ public class JakartaPersistenceTaskService implements TaskService {
 
     @Override
     public List<Task> getAllTasks() {
-        if (!(_securityContext.isCallerInRole("Sales")
-            || _securityContext.isCallerInRole("Shipping"))) {
-            throw new RuntimeException("Access denied. Role not allowed.");
+        // Anonymous users are not allowed
+        // Shipping role can view their own tasks
+        // Sales role can view all tasks
+        String username = _securityContext.getCallerPrincipal().getName();
+        if (username.equalsIgnoreCase("anonymous")) {
+            throw new SecurityException("Access denied. Anonymous users are not allowed.");
         }
-        if (_securityContext.isCallerInRole("Sales")) {
-            String username = _securityContext.getCallerPrincipal().getName();
-            return entityManager.createQuery("SELECT o FROM Task o where o.username = :usernameValue", Task.class)
-                    .setParameter("usernameValue", username)
+        boolean hasShippingRole = _securityContext.isCallerInRole("Shipping");
+        if (hasShippingRole) {
+            return entityManager.createQuery("SELECT o FROM Task o WHERE o.username = :uname", Task.class)
+                    .setParameter("uname", username)
                     .getResultList();
         }
-        return entityManager.createQuery("SELECT o FROM Task o ", Task.class)
-                .getResultList();
+        boolean hasSalesRole = _securityContext.isCallerInRole("Sales");
+        if (hasSalesRole) {
+            return entityManager.createQuery("SELECT o FROM Task o ", Task.class)
+                    .getResultList();
+        }
+        throw new SecurityException("Access denied. Your role does not allow you to access this resource.");
     }
 
     @Override
     @Transactional
-    public Task updateTask(@Valid Task task) {
+    public Task updateTask(Task task) {
 
         Optional<Task> optionalTask = getTaskById(task.getId());
         if (optionalTask.isEmpty()) {
@@ -81,6 +87,7 @@ public class JakartaPersistenceTaskService implements TaskService {
             throw new RuntimeException(errorMessage);
         } else {
             var existingTask = optionalTask.orElseThrow();
+            // Update only properties that is editable by the end user
 
             existingTask.setDescription(task.getDescription());
             existingTask.setPriority(task.getPriority());
@@ -97,6 +104,11 @@ public class JakartaPersistenceTaskService implements TaskService {
         Optional<Task> optionalTask = getTaskById(id);
         if (optionalTask.isPresent()) {
             Task task = optionalTask.orElseThrow();
+            // Authenticate users can only their own records
+            String username = _securityContext.getCallerPrincipal().getName();
+            if (!task.getUsername().equals(username)) {
+                throw new SecurityException("Access denied. You cannot delete data owned another user.");
+            }
             // Write code to throw a RuntimeException if this entity contains child records
             entityManager.remove(task);
         } else {
